Docker_Prophet/requirements.txt
datetime
python-dateutil
pandas
numpy
google-cloud-bigquery
google-cloud-storage
google-api-python-client
google
db-dtypes
prophet==1.1.5
optuna==3.5.0
dask==2024.1.1
sktime[all_extras]==0.25.0
distributed==2024.1.1